requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
ccxt==4.3.85
python-dotenv==1.0.1
redis==5.0.7
httpx==0.27.0
pydantic==2.8.2
